# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/monthly-quiz/current": {
      "get": {
        "operationId": "MonthlyQuizController_getCurrentMonthlyQuiz",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Enveloppe avec quiz ou null si un bilan (score_history) existe déjà pour le mois civil UTC courant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyQuizCurrentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Écran quiz du mois : sous-quiz éventuel + métadonnées (dernier score, prochaine ouverture, fraîcheur)",
        "tags": [
          "monthly-quiz"
        ]
      }
    },
    "/monthly-quiz/dev/simulate-next-month": {
      "post": {
        "operationId": "MonthlyQuizController_devSimulateNextMonth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "score_history du mois courant clonés en mois précédent puis supprimés ; lastMonthlyQuizAt recalculé.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "example": true
                    },
                    "scoreRowsShifted": {
                      "type": "number",
                      "example": 1
                    },
                    "monthlyQuizzesUpdated": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Aucune donnée du mois courant à simuler."
          },
          "403": {
            "description": "Desactive en production (NODE_ENV=production)."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[DEV uniquement] Simule le mois suivant : score_history et verrous du mois UTC courant sont déplacés vers le mois précédent.",
        "tags": [
          "monthly-quiz"
        ]
      }
    },
    "/monthly-quiz/{id}/submit": {
      "post": {
        "operationId": "MonthlyQuizController_submitMonthlyQuiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Identifiant du quiz (ex: quiz-1).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitMonthlyQuizDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Score carbone recalcule et enregistre dans l'historique utilisateur."
          },
          "409": {
            "description": "Quiz mensuel deja complete ce mois."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soumet les reponses du quiz mensuel et declenche le recalcul du score",
        "tags": [
          "monthly-quiz"
        ]
      }
    },
...
  },
  "info": {
    "title": "Projet-AGL API",
    "description": "The API documentation for Projet-AGL backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "MonthlyQuizOptionDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "example": "Voiture individuelle"
          },
          "value": {
            "type": "string",
            "example": "voiture_diesel_essence"
          },
          "description": {
            "type": "string",
            "example": "Le mode le plus emetteur."
          }
        },
        "required": [
          "label",
          "value"
        ]
      },
      "MonthlyQuizQuestionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "q-transport-mode"
          },
          "type": {
            "type": "string",
            "example": "single"
          },
          "title": {
            "type": "string",
            "example": "Comment vous rendez-vous au travail ?"
          },
          "description": {
            "type": "string",
            "example": "Le transport represente ~30 %."
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyQuizOptionDto"
            }
          },
          "min": {
            "type": "number",
            "example": 0
          },
          "max": {
            "type": "number",
            "example": 30
          },
          "showIf": {
            "type": "object"
          },
          "carbonMeta": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "type",
          "title"
        ]
      },
      "MonthlyQuizCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cat-transport"
          },
          "name": {
            "type": "string",
            "example": "MOBILITE & TRANSPORT"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyQuizQuestionDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "questions"
        ]
      },
      "MonthlyQuizResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "quiz-1"
          },
          "name": {
            "type": "string",
            "example": "Bilan carbone"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyQuizCategoryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "categories"
        ]
      },
      "MonthlyQuizCurrentDto": {
        "type": "object",
        "properties": {
          "quiz": {
            "nullable": true,
            "description": "Sous-quiz du mois (4 catégories) si une mise à jour est possible ce mois-ci, sinon null.",
            "allOf": [
              {
                "$ref": "#/components/schemas/MonthlyQuizResponseDto"
              }
            ]
          },
          "lastScoreHistoryAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Date ISO du dernier score enregistré (score_history.created_at).",
            "example": "2026-05-01T12:00:00.000Z"
          },
          "nextMonthlyQuizAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Prochaine ouverture prévue du quiz mensuel (users.next_monthly_quiz_at), renseignée après une soumission."
          },
          "dataFreshness": {
            "type": "string",
            "enum": [
              "none",
              "recent",
              "stale"
            ],
            "description": "none = aucun historique, recent = dernier score ≤ 21 jours, stale = dernier score > 21 jours."
          },
          "baselineAnswers": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Réponses du dernier enregistrement (score_history) pour préremplir l’UI et résoudre les showIf ; le POST submit n’envoie que les questions mises à jour."
          }
        },
        "required": [
          "quiz",
          "lastScoreHistoryAt",
          "nextMonthlyQuizAt",
          "dataFreshness",
          "baselineAnswers"
        ]
      },
      "SubmitMonthlyQuizDto": {
        "type": "object",
        "properties": {
          "answers": {
            "type": "object",
            "description": "Map des reponses du quiz mensuel. Cle = questionId, valeur = string | number | string[]",
            "example": {
              "q-transport-mode": "voiture_diesel_essence",
              "q-transport-distance": "50_100",
              "q-housing-surface": 45
            }
          }
        },
        "required": [
          "answers"
        ]
      },
...
    }
  }
}